+++ test_server/__init__.py
@@ -3,6 +3,7 @@ from docker.models.containers import Container
 from time import sleep
 from typing import cast
 import docker.errors
+import os
 
 
 def _getenvs():
@@ -62,21 +63,45 @@ def start_mssql_server() -> Container:
     return sql_server
 
 
+def get_test_blobstorage():
+    constr = os.getenv(
+        "TEST_BLOB_CONSTR",
+        "DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;",
+    )
+    from azure.storage.blob import ContainerClient
+
+    cc = ContainerClient.from_connection_string(constr, "testlake")
+    if not cc.exists():
+        cc.create_container()
+    return cc
+
+
+def upload_to_azurite():
+    with get_test_blobstorage() as cc:
+        faker_pq = "tests/data/startest/faker.parquet"
+
+        with open(faker_pq, "rb") as f:
+            cc.upload_blob("td/faker.parquet", f)
+        for root, _, fls in os.walk("tests/delta/fake"):
+            for fl in fls:
+                with open(os.path.join(root, fl), "rb") as f:
+                    cc.upload_blob(f"td/delta/fake/{fl}", f)
+
+
 def start_azurite() -> Container:
     client = docker.from_env()  # code taken from https://github.com/fsspec/adlfs/blob/main/adlfs/tests/conftest.py#L72
     azurite_server: Container | None = None
     try:
         m = cast(Container, client.containers.get("test4azurite"))
         if m.status == "running":
+            upload_to_azurite()
             return m
         else:
-            sql_server = m
+            azurite_server = m
     except docker.errors.NotFound as err:
         pass
 
     if azurite_server is None:
-        # using podman:  podman run  --env-file=TESTS/SQL_DOCKER.ENV --publish=1439:1433 --name=mssql1 chriseaton/adventureworks:light
-        #                podman kill mssql1
         azurite_server = client.containers.run(
             "mcr.microsoft.com/azure-storage/azurite",
             detach=True,
@@ -86,7 +111,7 @@ def start_azurite() -> Container:
     assert azurite_server is not None
     azurite_server.start()
     print(azurite_server.status)
-    sleep(45)  # the install script takes a sleep of 30s to create the db and then restores adventureworks
-
+    sleep(20)
+    upload_to_azurite()
     print("Successfully created azurite container...")
     return azurite_server
